Bump minimum PHP requirement to 7.4 and update PHPUnit
The code changes in the tests is to make the tests pass.

Also set the same namespace in tests and the tested code. This is not strictly necessary for the
tests to pass, but it makes testing somewhat more conveinent.

Will also try to run tests on PHP 8 in the workflow.